modules/vishnu-orchestrator/src/services/policy_engine.py
"""
🛡️ Vishnu Policy Engine
=======================

Advanced policy management and enforcement engine.
"""

import asyncio
import logging
from typing import Dict, List, Any, Optional
from datetime import datetime
import json
import yaml

from ..models.policy import (
    Policy, PolicyViolation, RemediationAction, PolicyStatus
)

logger = logging.getLogger(__name__)


class PolicyValidationResult:
    """Policy validation result"""
    def __init__(self, is_valid: bool, errors: List[str] = None):
        self.is_valid = is_valid
        self.errors = errors or []


class PolicyEngine:
    """
    Divine policy engine for continuous enforcement
    """

    def __init__(self):
        self.opa_client = None  # Would be initialized with OPA client
        self.violation_handlers = {}
        self.remediation_templates = {}
        self._load_remediation_templates()

    async def validate_policy_rules(self, rules: Dict[str, Any]) -> PolicyValidationResult:
        """
        Validate policy rules syntax and semantics
        """
        try:
            logger.info("Validating policy rules")
            
            # Basic validation
            if not isinstance(rules, dict):
                return PolicyValidationResult(False, ["Rules must be a dictionary"])
            
            required_fields = ["package", "default", "rules"]
            missing_fields = [f for f in required_fields if f not in rules]
            if missing_fields:
                return PolicyValidationResult(
                    False, 
                    [f"Missing required fields: {missing_fields}"]
                )
            
            # Validate OPA Rego syntax (simplified)
            if not isinstance(rules.get("package"), str):
                return PolicyValidationResult(
                    False, 
                    ["Package must be a string"]
                )
            
            # Would integrate with OPA server for full validation
            # For now, basic structure validation
            return PolicyValidationResult(True)
            
        except Exception as e:
            logger.error(f"Error validating policy rules: {str(e)}")
            return PolicyValidationResult(False, [str(e)])

    async def deploy_policy(self, policy: Policy) -> bool:
        """
        Deploy policy to enforcement engine
        """
        try:
            logger.info(f"Deploying policy {policy.id}")
            
            # Convert policy to OPA bundle
            opa_bundle = {
                "data": {
                    "policies": {
                        policy.id: {
                            "metadata": {
                                "name": policy.name,
                                "type": policy.policy_type,
                                "scope": policy.scope,
                                "enforcement_mode": policy.enforcement_mode
                            },
                            "rules": policy.rules
                        }
                    }
                }
            }
            
            # Deploy to OPA server (mock implementation)
            await self._deploy_to_opa(policy.id, opa_bundle)
            
            # Schedule continuous evaluation
            await self._schedule_policy_evaluation(policy)
            
            logger.info(f"Policy {policy.id} deployed successfully")
            return True
            
        except Exception as e:
            logger.error(f"Error deploying policy {policy.id}: {str(e)}")
            return False

    async def evaluate_policy(self, policy_id: str, 
                            resource_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Evaluate policy against resource data
        """
        try:
            logger.debug(f"Evaluating policy {policy_id}")
            
            # Get policy
            policy = await Policy.get(policy_id)
            if not policy:
                raise ValueError(f"Policy {policy_id} not found")
            
            # Prepare evaluation input
            input_data = {
                "resource": resource_data,
                "metadata": {
                    "timestamp": datetime.utcnow().isoformat(),
                    "policy_id": policy_id
                }
            }
            
            # Evaluate against OPA
            result = await self._evaluate_with_opa(policy.rules, input_data)
            
            # Process violations
            if result.get("violations"):
                await self._handle_violations(policy, result["violations"])
            
            return result
            
        except Exception as e:
            logger.error(f"Error evaluating policy {policy_id}: {str(e)}")
            return {"error": str(e)}

    async def remediate_violation(self, violation: PolicyViolation) -> Dict[str, Any]:
        """
        Execute automated remediation for policy violation
        """
        try:
            logger.info(f"Remediating violation {violation.id}")
            
            # Get remediation template
            template = self.remediation_templates.get(violation.violation_type)
            if not template:
                return {
                    "success": False,
                    "error": f"No remediation template for {violation.violation_type}"
                }
            
            # Create remediation action
            action = RemediationAction(
                violation_id=violation.id,
                action_type=template["type"],
                action_details=template["details"],
                triggered_by="system"
            )
            
            # Execute remediation
            result = await self._execute_remediation(action, violation)
            
            # Update violation status
            if result["success"]:
                violation.status = "remediated"
                violation.remediated_at = datetime.utcnow()
                violation.remediation_action = result["action_summary"]
                # await violation.save()
            
            return result
            
        except Exception as e:
            logger.error(f"Error remediating violation {violation.id}: {str(e)}")
            return {"success": False, "error": str(e)}

    async def _deploy_to_opa(self, policy_id: str, bundle: Dict[str, Any]):
        """Deploy policy bundle to OPA server"""
        # Mock implementation - would use OPA client
        logger.debug(f"Deploying bundle for policy {policy_id} to OPA")
        await asyncio.sleep(0.1)  # Simulate network call

    async def _schedule_policy_evaluation(self, policy: Policy):
        """Schedule continuous policy evaluation"""
        # Would integrate with scheduler (Celery, APScheduler, etc.)
        logger.debug(f"Scheduling evaluation for policy {policy.id}")

    async def _evaluate_with_opa(self, rules: Dict[str, Any], 
                                input_data: Dict[str, Any]) -> Dict[str, Any]:
        """Evaluate rules with OPA engine"""
        # Mock implementation - would use OPA client
        logger.debug("Evaluating with OPA engine")
        
        # Simulate policy evaluation
        await asyncio.sleep(0.1)
        
        # Mock result based on input
        if "security" in str(rules).lower():
            return {
                "allow": True,
                "violations": [],
                "score": 95.0,
                "details": "Security policy compliance verified"
            }
        
        return {
            "allow": True,
            "violations": [],
            "score": 100.0,
            "details": "Policy evaluation completed"
        }

    async def _handle_violations(self, policy: Policy, violations: List[Dict]):
        """Handle policy violations"""
        for violation_data in violations:
            violation = PolicyViolation(
                policy_id=policy.id,
                resource_id=violation_data.get("resource_id"),
                violation_type=violation_data.get("type"),
                severity=violation_data.get("severity", "medium"),
                violation_details=violation_data
            )
            # await violation.save()
            
            # Trigger automated remediation if enabled
            if policy.enforcement_mode == "enforce":
                await self.remediate_violation(violation)

    async def _execute_remediation(self, action: RemediationAction, 
                                 violation: PolicyViolation) -> Dict[str, Any]:
        """Execute remediation action"""
        try:
            logger.info(f"Executing remediation action {action.action_type}")
            
            # Mock implementation - would execute actual remediation
            action.started_at = datetime.utcnow()
            
            # Simulate execution
            await asyncio.sleep(1)
            
            action.status = "completed"
            action.completed_at = datetime.utcnow()
            action.result = {
                "success": True,
                "changes_applied": ["Mock remediation executed"],
                "verification": "Policy compliance restored"
            }
            
            return {
                "success": True,
                "action_summary": f"Applied {action.action_type} remediation",
                "details": action.result
            }
            
        except Exception as e:
            action.status = "failed"
            action.error_message = str(e)
            return {"success": False, "error": str(e)}

    def _load_remediation_templates(self):
        """Load remediation templates"""
        self.remediation_templates = {
            "security_group_violation": {
                "type": "modify_security_group",
                "details": {
                    "action": "remove_open_ports",
                    "ports": ["22", "3389", "443"]
                }
            },
            "encryption_violation": {
                "type": "enable_encryption",
                "details": {
                    "action": "enable_encryption_at_rest",
                    "algorithm": "AES-256"
                }
            },
            "cost_violation": {
                "type": "resource_optimization",
                "details": {
                    "action": "right_size_instance",
                    "recommendation": "downgrade_instance_type"
                }
            }
        }
