modules/vishnu-orchestrator/requirements.txt
# 🛡️ Vishnu Orchestrator Engine Dependencies
# ==========================================

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Kubernetes integration
kubernetes==28.1.0
kubernetes-asyncio==28.2.0

# Policy engines
py-opa==0.8.0
casbin==1.33.0
casbin-sqlalchemy-adapter==1.4.0

# Workflow orchestration
argo-workflows==6.4.8
celery==5.3.4
kombu==5.3.4

# Service mesh & networking
istio-client-python==1.19.0

# Monitoring & observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Cloud providers
azure-identity==1.15.0
azure-mgmt-containerservice==29.2.0
boto3==1.34.0
google-cloud-container==2.32.0

# Configuration management
python-dotenv==1.0.0
pyyaml==6.0.1

# Caching
redis==5.0.1
aioredis==2.0.1

# Security
cryptography==41.0.8
pyjwt==2.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
